Create behavior for repositories

+++ src/main/java/project/domain/sweet/Bakery.java
@@ -7,7 +7,7 @@ import static project.domain.sweet.choice.SweetType.BAKERY;
 public abstract class Bakery extends AbstractSweet {
     private static final SweetType TYPE = BAKERY;
 
-    public Bakery(String name, Long price, Integer weight, Byte sugarContent) {
+    public Bakery(String name, Integer price, Integer weight, Byte sugarContent) {
         super(name, price, weight, sugarContent);
     }
