+++ src/main/java/project/domain/sweet/Chocolate.java
@@ -9,7 +9,7 @@ import static project.domain.sweet.choice.SweetType.CHOCOLATE;
 public abstract class Chocolate extends AbstractSweet {
     private static final SweetType TYPE = CHOCOLATE;
 
-    public Chocolate(String name, Long price, Integer weight, Byte sugarContent) {
+    public Chocolate(String name, Integer price, Integer weight, Byte sugarContent) {
         super(name, price, weight, sugarContent);
     }
